Deconstruct into individual targets
As support for older browsers is dropped, we can slowly remove some of
these shims to reduce the amount of JavaScript we ship. So this can be
granular, I've split up the main entry point into separate targets. With
this organization, you simply pick the oldest version that you need to
shim with the default being ES5.