Remove webrtc-adapter
I am working to reduce the amount of JavaScript that we ship on most
pages of airbnb.com. One package that I noticed was rather large and
shipped on every page in a way that runs before all of the application
code is the webrtc-adapter package. This is a shim to insulate apps from
spec changes and prefix differences. However, we don't use WebRTC on
most of our pages, so I am moving it out of our browser shims package
and into the specific places that need it.

This is a breaking change.